Fix code : - Rename and change folder for models. - Remove useless code
