+++ OOS/package.json
@@ -23,7 +23,7 @@
     "@angular/router": "^5.2.0",
     "angular-font-awesome": "^3.1.2",
     "angular-in-memory-web-api": "^0.5.4",
-    "bootstrap": "^4.0.0",
+    "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "jquery": "^3.3.1",
